new dynamic sampler test

+++ mmengine/dataset/sampler.py
@@ -210,7 +210,7 @@ class DynamicSampler(Sampler):
 
         # get indices of elements which should be included in training 
         counts = torch.zeros(self.num_classes, dtype = torch.int)
-        indices = torch.empty(self.sample_size.sum(), dtype=torch.long)
+        indices = torch.empty((self.sample_size.int().sum(), ), dtype=torch.long)
         start_idx = 0
         if self.enable_ROS:
             for label in range(0, self.num_classes):
